+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.4
 
 - Fix: Private super class can now be exported to child classes.
+- Breaking Changes: Adapt to the Hetu script changes. 
 
 ## 1.0.3